Move the scripts/mongo/foo2.php scripts to foo.php
